refactor: extract conversation components to separate files
This refactoring breaks down the large conversation.tsx file (569 lines) into smaller, focused components:

- CopyLinkButton.tsx (28 lines)
- ScrollToTopButton.tsx (62 lines)
- MessageThreadPanel.tsx (32 lines)
- MessageActionsPanel.tsx (12 lines)
- ConversationHeader.tsx (68 lines)
- ErrorContent.tsx (17 lines)
- LoadingContent.tsx (9 lines)
- CarouselPreviewContent.tsx (54 lines)
- MergedContent.tsx (12 lines)

This reduces the main conversation.tsx file from 569 lines to 239 lines (330 line reduction)
while maintaining exact functionality and improving code organization.

Benefits:
- Better separation of concerns
- Improved maintainability
- Easier testing of individual components
- Reduced cognitive load for developers
