+++ app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/conversation.tsx
@@ -1,27 +1,10 @@
-import FileSaver from "file-saver";
-import {
-  ArrowUp,
-  ChevronLeft,
-  ChevronRight,
-  Download,
-  Info,
-  Link as LinkIcon,
-  PanelRightClose,
-  PanelRightOpen,
-  X,
-} from "lucide-react";
-import Link from "next/link";
-import { useEffect, useLayoutEffect, useState } from "react";
-import { useMediaQuery } from "react-responsive";
+import { useLayoutEffect, useState } from "react";
 import { useStickToBottom } from "use-stick-to-bottom";
 import {
   ConversationContextProvider,
   useConversationContext,
 } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/conversationContext";
-import { MessageThread } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/messageThread";
-import Viewers from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/conversation/viewers";
 import { useConversationListContext } from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/list/conversationListContext";
-import PreviewModal from "@/app/(dashboard)/mailboxes/[mailbox_slug]/[category]/previewModal";
 import {
   type AttachedFile,
   type ConversationEvent,
@@ -29,12 +12,7 @@ import {
   type Message,
   type Note,
 } from "@/app/types/global";
-import { CarouselDirection, createCarousel } from "@/components/carousel";
-import LoadingSpinner from "@/components/loadingSpinner";
-import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Dialog, DialogContent, DialogHeader, DialogTitle } from "@/components/ui/dialog";
 import { ResizableHandle, ResizablePanel, ResizablePanelGroup } from "@/components/ui/resizable";
 import { SidebarProvider } from "@/components/ui/sidebar";
 import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
@@ -46,324 +24,16 @@ import { cn } from "@/lib/utils";
 import { api } from "@/trpc/react";
 import { useConversationsListInput } from "../shared/queries";
 import ConversationSidebar from "./conversationSidebar";
-import { MessageActions } from "./messageActions";
+import { CarouselPreviewContent } from "@/components/conversation/CarouselPreviewContent";
+import { ConversationHeader } from "@/components/conversation/ConversationHeader";
+import { ErrorContent } from "@/components/conversation/ErrorContent";
+import { LoadingContent } from "@/components/conversation/LoadingContent";
+import { MergedContent } from "@/components/conversation/MergedContent";
+import { MessageActionsPanel } from "@/components/conversation/MessageActionsPanel";
+import { MessageThreadPanel } from "@/components/conversation/MessageThreadPanel";
 
 export type ConversationWithNewMessages = Omit<ConversationType, "messages"> & {
   messages: ((Message | Note | ConversationEvent) & { isNew?: boolean })[];
-};
-
-const { Carousel, CarouselButton, CarouselContext } = createCarousel<AttachedFile>();
-
-const CopyLinkButton = () => {
-  const isStandalone = useMediaQuery({ query: "(display-mode: standalone)" });
-  const [copied, setCopied] = useState(false);
-
-  if (!isStandalone) return null;
-
-  return (
-    <Tooltip delayDuration={0}>
-      <TooltipTrigger asChild>
-        <Button
-          variant="ghost"
-          size="sm"
-          iconOnly
-          onClick={async (e) => {
-            e.preventDefault();
-            const url = window.location.href;
-            await navigator.clipboard.writeText(url);
-            setCopied(true);
-            setTimeout(() => setCopied(false), 2000);
-          }}
-        >
-          <LinkIcon className="h-4 w-4" />
-          <span className="sr-only">Copy link</span>
-        </Button>
-      </TooltipTrigger>
-      <TooltipContent>{copied ? "Copied!" : "Copy link"}</TooltipContent>
-    </Tooltip>
-  );
-};
-
-const ScrollToTopButton = ({
-  scrollRef,
-}: {
-  scrollRef: React.MutableRefObject<HTMLElement | null> & React.RefCallback<HTMLElement>;
-}) => {
-  const [show, setShow] = useState(false);
-
-  useEffect(() => {
-    const scrollElement = scrollRef.current;
-    if (!scrollElement) return;
-
-    let timeoutId: NodeJS.Timeout;
-    const handleScroll = () => {
-      if (timeoutId) {
-        clearTimeout(timeoutId);
-      }
-
-      timeoutId = setTimeout(() => {
-        const scrollTop = scrollElement.scrollTop;
-        const threshold = 100;
-
-        // Show button whenever scrolled past threshold
-        setShow(scrollTop > threshold);
-      }, 100);
-    };
-
-    scrollElement.addEventListener("scroll", handleScroll);
-    return () => {
-      scrollElement.removeEventListener("scroll", handleScroll);
-      if (timeoutId) {
-        clearTimeout(timeoutId);
-      }
-    };
-  }, [scrollRef]);
-
-  const scrollToTop = () => {
-    scrollRef.current?.scrollTo({
-      top: 0,
-      behavior: "smooth",
-    });
-  };
-
-  return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <a
-          className={cn(
-            "absolute bottom-4 left-4 transition-all duration-200 h-8 w-8 p-0 rounded-full",
-            "flex items-center justify-center",
-            "bg-background border border-border shadow-xs",
-            "hover:border-primary hover:shadow-md hover:bg-muted",
-            show ? "translate-y-0 opacity-100" : "translate-y-4 opacity-0 pointer-events-none",
-          )}
-          onClick={scrollToTop}
-          aria-label="Scroll to top"
-        >
-          <ArrowUp className="h-4 w-4 text-foreground" />
-        </a>
-      </TooltipTrigger>
-      <TooltipContent>Scroll to top</TooltipContent>
-    </Tooltip>
-  );
-};
-
-const MessageThreadPanel = ({
-  scrollRef,
-  contentRef,
-  setPreviewFileIndex,
-  setPreviewFiles,
-}: {
-  scrollRef: React.MutableRefObject<HTMLElement | null> & React.RefCallback<HTMLElement>;
-  contentRef: React.MutableRefObject<HTMLElement | null>;
-  setPreviewFileIndex: (index: number) => void;
-  setPreviewFiles: (files: AttachedFile[]) => void;
-}) => {
-  const { mailboxSlug, data: conversationInfo } = useConversationContext();
-
-  return (
-    <div className="grow overflow-y-auto relative" ref={scrollRef}>
-      <div ref={contentRef as React.RefObject<HTMLDivElement>} className="relative">
-        <ScrollToTopButton scrollRef={scrollRef} />
-        <div className="flex flex-col gap-8 px-4 py-4 h-full">
-          {conversationInfo && (
-            <MessageThread
-              mailboxSlug={mailboxSlug}
-              conversation={conversationInfo}
-              onPreviewAttachment={(message, currentIndex) => {
-                setPreviewFileIndex(currentIndex);
-                setPreviewFiles(message.files);
-              }}
-            />
-          )}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const MessageActionsPanel = () => {
-  return (
-    <div
-      className="h-full bg-muted px-4 pb-4"
-      onKeyDown={(e) => {
-        // Prevent keypress events from triggering the global inbox view keyboard shortcuts
-        e.stopPropagation();
-      }}
-    >
-      <MessageActions />
-    </div>
-  );
-};
-
-const ConversationHeader = ({
-  conversationMetadata,
-  isAboveSm,
-  sidebarVisible,
-  setSidebarVisible,
-}: {
-  conversationMetadata: any;
-  isAboveSm: boolean;
-  sidebarVisible: boolean;
-  setSidebarVisible: (visible: boolean) => void;
-}) => {
-  const { mailboxSlug, data: conversationInfo } = useConversationContext();
-  const { minimize, moveToNextConversation, moveToPreviousConversation, currentIndex, currentTotal, hasNextPage } =
-    useConversationListContext();
-
-  return (
-    <div
-      className={cn(
-        "flex items-center border-b border-border h-12 px-2 md:px-4 gap-x-2",
-        !conversationInfo && "hidden",
-      )}
-      style={{ minHeight: 48 }}
-    >
-      <div className="flex items-center min-w-0 flex-shrink-0 z-10 lg:w-44">
-        <Button variant="ghost" size="sm" iconOnly onClick={minimize} className="text-primary hover:text-foreground">
-          <X className="h-4 w-4" />
-        </Button>
-        <div className="flex items-center ml-2">
-          <Button variant="ghost" size="sm" iconOnly onClick={moveToPreviousConversation}>
-            <ChevronLeft className="h-4 w-4" />
-          </Button>
-          <span className="text-sm text-muted-foreground whitespace-nowrap text-center mx-1">
-            {currentIndex + 1} of {currentTotal}
-            {hasNextPage ? "+" : ""}
-          </span>
-          <Button variant="ghost" size="sm" iconOnly onClick={moveToNextConversation}>
-            <ChevronRight className="h-4 w-4" />
-          </Button>
-        </div>
-      </div>
-      <div className="flex-1 min-w-0 flex justify-center">
-        <div className="truncate text-base font-semibold text-foreground text-center max-w-full">
-          {conversationMetadata.subject ?? "(no subject)"}
-        </div>
-      </div>
-      <div className="flex items-center gap-2 min-w-0 flex-shrink-0 z-10 lg:w-44 justify-end">
-        <CopyLinkButton />
-        {conversationInfo?.id && <Viewers mailboxSlug={mailboxSlug} conversationSlug={conversationInfo.slug} />}
-        <Button
-          variant={!isAboveSm && sidebarVisible ? "subtle" : "ghost"}
-          size="sm"
-          iconOnly
-          onClick={() => setSidebarVisible(!sidebarVisible)}
-        >
-          {isAboveSm ? (
-            sidebarVisible ? (
-              <PanelRightClose className="h-4 w-4" />
-            ) : (
-              <PanelRightOpen className="h-4 w-4" />
-            )
-          ) : (
-            <Info className="h-4 w-4" />
-          )}
-          <span className="sr-only">{sidebarVisible ? "Hide sidebar" : "Show sidebar"}</span>
-        </Button>
-      </div>
-    </div>
-  );
-};
-
-const ErrorContent = () => {
-  const { error, refetch } = useConversationContext();
-  if (!error) return null;
-
-  return (
-    <div className="flex items-center justify-center grow">
-      <Alert variant="destructive" className="max-w-lg text-center">
-        <AlertTitle>Failed to load conversation</AlertTitle>
-        <AlertDescription className="flex flex-col gap-4">
-          Error loading this conversation: {error.message}
-          <Button variant="destructive_outlined" onClick={() => refetch()}>
-            Try again
-          </Button>
-        </AlertDescription>
-      </Alert>
-    </div>
-  );
-};
-
-const LoadingContent = () => {
-  const { isPending } = useConversationContext();
-  if (!isPending) return null;
-
-  return (
-    <div className="flex items-center justify-center grow">
-      <LoadingSpinner size="md" />
-    </div>
-  );
-};
-
-const CarouselPreviewContent = ({
-  previewFileIndex,
-  setPreviewFileIndex,
-  previewFiles,
-  setPreviewFiles,
-}: {
-  previewFileIndex: number;
-  setPreviewFileIndex: (index: number) => void;
-  previewFiles: AttachedFile[];
-  setPreviewFiles: (files: AttachedFile[]) => void;
-}) => {
-  return (
-    <CarouselContext.Provider
-      value={{
-        currentIndex: previewFileIndex,
-        setCurrentIndex: setPreviewFileIndex,
-        items: previewFiles,
-      }}
-    >
-      <Carousel>
-        {(currentFile) => (
-          <Dialog open={!!currentFile} onOpenChange={(open) => !open && setPreviewFiles([])}>
-            <DialogContent className="max-w-5xl">
-              <DialogHeader>
-                <DialogTitle>File Preview</DialogTitle>
-              </DialogHeader>
-              <div className="relative bottom-0.5 flex items-center justify-between p-3">
-                <div className="max-w-xs truncate" title={currentFile.name}>
-                  {currentFile.name}
-                </div>
-
-                <div className="mr-6 flex items-center">
-                  <button
-                    onClick={() =>
-                      currentFile.presignedUrl && FileSaver.saveAs(currentFile.presignedUrl, currentFile.name)
-                    }
-                  >
-                    <Download className="text-primary h-5 w-5 shrink-0" />
-                    <span className="sr-only">Download</span>
-                  </button>
-                </div>
-              </div>
-
-              <div className="relative flex flex-row items-center justify-center gap-3">
-                <CarouselButton direction={CarouselDirection.LEFT} className="absolute -left-10 md:-left-11" />
-                <PreviewModal file={currentFile} />
-                <CarouselButton direction={CarouselDirection.RIGHT} className="absolute -right-10 md:-right-11" />
-              </div>
-            </DialogContent>
-          </Dialog>
-        )}
-      </Carousel>
-    </CarouselContext.Provider>
-  );
-};
-
-const MergedContent = () => {
-  const { mailboxSlug, data: conversationInfo } = useConversationContext();
-  if (!conversationInfo?.mergedInto?.slug) return null;
-
-  return (
-    <div className="absolute inset-0 z-50 bg-background/75 flex flex-col items-center justify-center gap-4 h-full text-lg">
-      Merged into another conversation.
-      <Button variant="subtle" asChild>
-        <Link href={`/mailboxes/${mailboxSlug}/conversations?id=${conversationInfo.mergedInto.slug}`}>View</Link>
-      </Button>
-    </div>
-  );
 };
 
 const ConversationContent = () => {
