+++ components/conversation/CopyLinkButton.tsx
@@ -1,0 +1,35 @@
+import { LinkIcon } from "lucide-react";
+import { useState } from "react";
+import { useMediaQuery } from "react-responsive";
+import { Button } from "@/components/ui/button";
+import { Tooltip, TooltipContent, TooltipTrigger } from "@/components/ui/tooltip";
+
+export const CopyLinkButton = () => {
+  const isStandalone = useMediaQuery({ query: "(display-mode: standalone)" });
+  const [copied, setCopied] = useState(false);
+
+  if (!isStandalone) return null;
+
+  return (
+    <Tooltip delayDuration={0}>
+      <TooltipTrigger asChild>
+        <Button
+          variant="ghost"
+          size="sm"
+          iconOnly
+          onClick={async (e) => {
+            e.preventDefault();
+            const url = window.location.href;
+            await navigator.clipboard.writeText(url);
+            setCopied(true);
+            setTimeout(() => setCopied(false), 2000);
+          }}
+        >
+          <LinkIcon className="h-4 w-4" />
+          <span className="sr-only">Copy link</span>
+        </Button>
+      </TooltipTrigger>
+      <TooltipContent>{copied ? "Copied!" : "Copy link"}</TooltipContent>
+    </Tooltip>
+  );
+};
